A small amount of cleanup and work on FT Sportster lights. Still need to integrate all the modes.
Started a test project to try and get ESC Telemetry from a KISS enabled esc.
So far no luck, I've checked timings for DSHOT 300 with an oscilliscope and they seem good.  I kind of suspect I need to do something to prep the ESC.  Anyway still a work in progress.