Huge rewrite of dipper lights control for new build of dual polaris plane. This also has a much better host/client managements system to keep the two controllers in sync between the two hulls. This could be used by other projects to expand the # of PWM by simply having additional controller boards hooked to a host controller.
Several of the components in this project will probably be split out into utility libraries:
SynchableChannel which is a really simple packet based protocol used between the host/client that can resync if data is lost/corrupted.
LedSpan which provides easier access to a span of leds in particular it can reverse indexing, allows code to map the span of leds into a larger byte array, and provides some convenience functions for setting all of the leds in the span.  This was really useful for dipper since the logical order of the leds is slightly different in the two hulls.